added calendar input with dojo calendar popup. this includes proper validation message placement.
changed dijit theme to claro.

fixed some issues with validation messages having quotes in the text. this was messing up the json and html because an input element was being used. i changed it to a textarea so quotes won't mess up the html. the json can have double quotes now, but need to look into escaping single/double quotes to fully eliminate the problem.

added calendar to event and price pages.

changed required validation message for checkboxes/radios to a fixed string. this is preemptive since the label may be prohibitively long to use.